+++ src/main/java/com/ohacd/matchbox/api/events/GameStartEvent.java
@@ -1,9 +1,11 @@
 package com.ohacd.matchbox.api.events;
 
+import com.google.common.annotations.Beta;
 import com.ohacd.matchbox.api.MatchboxEvent;
 import com.ohacd.matchbox.api.MatchboxEventListener;
 import com.ohacd.matchbox.game.utils.Role;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
 import java.util.Map;
@@ -12,7 +14,7 @@ import java.util.Map;
  * Event fired when a new game starts.
  * 
  * @since 0.9.5
- * @author Matchbox Team
+ * @author OhACD
  */
 public class GameStartEvent extends MatchboxEvent {
     
@@ -27,14 +29,14 @@ public class GameStartEvent extends MatchboxEvent {
      * @param players all players in the game
      * @param roleAssignments mapping of players to their roles
      */
-    public GameStartEvent(String sessionName, Collection<Player> players, Map<Player, Role> roleAssignments) {
+    public GameStartEvent(@NotNull String sessionName, @NotNull Collection<Player> players, @NotNull Map<Player, Role> roleAssignments) {
         this.sessionName = sessionName;
         this.players = players;
         this.roleAssignments = roleAssignments;
     }
     
     @Override
-    public void dispatch(MatchboxEventListener listener) {
+    public void dispatch(@NotNull MatchboxEventListener listener) {
         listener.onGameStart(this);
     }
     
@@ -43,6 +45,7 @@ public class GameStartEvent extends MatchboxEvent {
      * 
      * @return the session name
      */
+    @NotNull
     public String getSessionName() {
         return sessionName;
     }
@@ -52,6 +55,7 @@ public class GameStartEvent extends MatchboxEvent {
      * 
      * @return collection of all players
      */
+    @NotNull
     public Collection<Player> getPlayers() {
         return players;
     }
@@ -61,6 +65,7 @@ public class GameStartEvent extends MatchboxEvent {
      * 
      * @return mapping of players to their assigned roles
      */
+    @NotNull
     public Map<Player, Role> getRoleAssignments() {
         return roleAssignments;
     }